# Fix local variable 'image_features' referenced before assignment