# What does this PR do?

PR fixes the following tests:
```
pytest -v -s tests/test_image_to_text_example.py -k llava-1.5
```
Failure:
```
UnboundLocalError: local variable 'image_features' referenced before assignment
```

Issue comes from transformers: https://github.com/huggingface/transformers/commit/9c4639b62236b503d881c225246ad8ee61ae6ed8#diff-2204a446d82a75771bf0df20c493c6b6cdfba7a87e51767ac63aa9cf8832100aR555
`image_features` is set inside an if statement  and the tests fails as they don't enter it.

#### Verify outputs/functional accuracy test
1. llava-1.5-7b-hf, BF16
```
python3 /root/synopsis/transformers_future/examples/image-to-text/run_pipeline.py --model_name_or_path llava-hf/llava-1.5-7b-hf --batch_size 1 --max_new_tokens 20 --use_hpu_graphs --bf16

# With this PR on transformers_future
09/30/2024 20:11:57 - INFO - __main__ - result = [[{'generated_text': 'USER:  \nWhat is shown in this image? ASSISTANT: The image shows a pier extending out over a body of water, with a mountain in the background.'}]]
09/30/2024 20:11:57 - INFO - __main__ - time = 224.79350978974253ms, Throughput (including tokenization) = 80.0734861822125 tokens/second
# On main
09/30/2024 20:16:39 - INFO - __main__ - result = [[{'generated_text': 'USER:  \nWhat is shown in this image? ASSISTANT: The image shows a pier extending out over a body of water, with a mountain in the background.'}]]
09/30/2024 20:16:39 - INFO - __main__ - time = 223.51140959654003ms, Throughput (including tokenization) = 80.53280158042831 tokens/second
```
2. llava-1.5-13b-hf, BF16
```
python3 /root/synopsis/transformers_future/examples/image-to-text/run_pipeline.py --model_name_or_path llava-hf/llava-1.5-13b-hf --batch_size 1 --max_new_tokens 20 --use_hpu_graphs --bf16
# With this PR on transformers_future
09/30/2024 20:13:36 - INFO - __main__ - result = [[{'generated_text': 'USER:  \nWhat is shown in this image? ASSISTANT: The image shows a wooden pier extending out into a large body of water, with mountains in the background'}]]
09/30/2024 20:13:36 - INFO - __main__ - time = 366.76659581717104ms, Throughput (including tokenization) = 49.07753379201631 tokens/second
# On main
09/30/2024 20:17:22 - INFO - __main__ - result = [[{'generated_text': 'USER:  \nWhat is shown in this image? ASSISTANT: The image shows a wooden pier extending out into a large body of water, with mountains in the background'}]]
09/30/2024 20:17:22 - INFO - __main__ - time = 364.24360398668796ms, Throughput (including tokenization) = 49.41747721301881 tokens/second
```
#### Regressions on transformers_future + this PR
```
pytest -v -s tests/test_image_to_text_example.py 
=10 passed in 665.23s (0:11:05) =
```

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

Fixes # (issue)


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?